feat

+++ device/lib/wav_file/src/WAVFileWriter.cpp
@@ -14,12 +14,12 @@ WAVFileWriter::WAVFileWriter(File fp, int sample_rate)
   m_file_size = sizeof(wav_header_t);
 }
 
-void WAVFileWriter::write(uint8_t *samples, int count)
+void WAVFileWriter::write(int16_t *samples, int count)
 {
   // write the samples and keep track of the file size so far
   m_fp.write(samples, count);
   //fwrite(samples, sizeof(int16_t), count, m_fp);
-  m_file_size += sizeof(uint8_t) * count;
+  m_file_size += sizeof(int16_t) * count;
 }
 
 void WAVFileWriter::finish()
